Move BufferSet deps to common
