+++ src/BufferSet.ts
@@ -3,10 +3,9 @@
  * @license MIT
  */
 
-import { IBufferSet } from './Types';
-import { IBuffer } from 'common/buffer/Types';
+import { IBuffer, IBufferSet } from 'common/buffer/Types';
 import { IAttributeData } from 'common/Types';
-import { Buffer } from './common/buffer/Buffer';
+import { Buffer } from 'common/buffer/Buffer';
 import { EventEmitter2, IEvent } from 'common/EventEmitter2';
 import { IOptionsService, IBufferService } from 'common/services/Services';
